Display Total Owed And Individual Bill Totals
Index view is now passed two dictionaries using Housemates as keys.

total_owed is a dictionary with values representing the total owed per
for that particular housemate.

individual_bills is a dictionary containing a list of all unpaid bills
so that a categorical breakdown of the owed total can be shown per
housemate.

Signed-off-by: Avery Uslaner <ctsadmin@conservationtechnologysolutions.org>